Move some stuff out into a base class
This will make it easier to write custom Stations only import
scripts (for example, if we get hold of CSVs with grid refs)
